x11/xdialog: register bsdconfig as an application
Since bsdconfig may be presumed as always available, and uses xdialog
when invoked for use as a graphical interface (ie with -X), this package
may just as well expose bsdconfig to the end user. This also installs a
scaled-down copy of the FreeBSD logo from /boot/images for use as its
icon (48 pixels tall and wide), where the trademark sign was removed
since it is unrecognizable as this scale. Finally, this also documents
the requirements in pkg-message. (Running as root or being a sudoer)

Sponsored-By: The FreeBSD Foundation