*/deforaos-*: take over maintainership
I am the original author of the DeforaOS desktop environment, that has
been packaged and maintained by olivier@ all these years. (Thanks!)

With his knowledge and agreement I am offering to be listed as the
official maintainer of these packages in FreeBSD, and will of course
welcome his assistance and supervision in this process.